Add Swagger,tests and some refactoring

+++ src/main/java/fi/korpi/bookapi/BookService.java
@@ -27,18 +27,30 @@ public class BookService {
     }
 
     public void deleteBook(Long id) {
+        Optional<Book> book = bookRepository.findById(id);
+        if (book.isEmpty()) {
+            throw new IllegalArgumentException("Book not found");
+        }
         bookRepository.deleteById(id);
     }
 
-    public List<Book> filterBooksByAuthor(String author) {
-        return bookRepository.findByAuthor(author);
-    }
-
-    public List<Book> filterBooksByYear(Integer year) {
-        return bookRepository.findByYear(year);
-    }
-
-    public List<Book> filterBooksByPublisher(String publisher) {
-        return bookRepository.findByPublisher(publisher);
+    public List<Book> filterBooks(String author, Integer year, String publisher) {
+        if (author != null && year != null && publisher != null) {
+            return bookRepository.findByAuthorAndYearAndPublisher(author, year, publisher);
+        } else if (author != null && year != null) {
+            return bookRepository.findByAuthorAndYear(author, year);
+        } else if (author != null && publisher != null) {
+            return bookRepository.findByAuthorAndPublisher(author, publisher);
+        } else if (year != null && publisher != null) {
+            return bookRepository.findByYearAndPublisher(year, publisher);
+        } else if (author != null) {
+            return bookRepository.findByAuthor(author);
+        } else if (year != null) {
+            return bookRepository.findByYear(year);
+        } else if (publisher != null) {
+            return bookRepository.findByPublisher(publisher);
+        } else {
+            return getAllBooks();
+        }
     }
 }
